% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\title{A squared value object}
\description{
Object of class \code{Candidate} as created by the \code{Candidate} functions
}
\details{
An object of the class `Candidate' has the following slots:
\itemize{
\item \code{name} Name of the Candidate
\item \code{delegatesWon} Delegates the Candidate has won so far
\item \code{party} The political party of the Candidate
\item \code{delegatesNeeded} How many more delegates the Candidate needs to win
}
}
\author{
David P. Flasterstein: \email{davidflasterstein@wustl.edu}
}

