% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCandidate.R
\name{createCandidate}
\alias{createCandidate}
\title{Create objects of class candidate}
\usage{
createCandidate(name, delegatesWon, party)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}
}
\value{
An object of class Candidate containing
 \item{name}{the name of the candidate}
 \item{delegatesWon}{The number of delegates won} 
 \item{party}{The political party of the candidate}
}
\description{
Creates an object of class candidate using some inputs
}
\examples{

createCandidate(name = "Hillary", delegatesWon = 100, party = "democrat")
}
\author{
David Flaserstein
}
\seealso{
\code{\link{Candidate}}
}

