\name{primaryRace-package}
\alias{primaryRace-package}
\alias{primaryRace}
\docType{package}
\title{
\packageTitle{primaryRace}
}
\description{
\packageDescription{primaryRace}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{primaryRace}
\packageIndices{primaryRace}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{primaryRace}

Maintainer: \packageMaintainer{primaryRace}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
