% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCandidate.R, R/propNeeded.R, R/propNeed.R
\docType{methods}
\name{print,Candidate-method}
\alias{print,Candidate-method}
\alias{propNeeded}
\title{Show objects of class Candidate}
\usage{
\S4method{print}{Candidate}(x, ...)

propNeeded(object = "Candidate", remainingDelegates = "numeric")

propNeeded(object = "Candidate", remainingDelegates = "numeric")
}
\arguments{
\item{x}{object of class Candidate}

\item{object}{of class Candidate}

\item{remainingDelegates}{number of remaining delegates}

\item{object}{of class Candidate}

\item{remainingDelegates}{number of remaining delegates}
}
\value{
print each individual element of Candidate

print each individual element of Candidate

print each individual element of Candidate
}
\description{
Inputing an object of class Candidate shows the object

Shows the porportion of remaining delegates candidate needs to win

Shows the porportion of remaining delegates candidate needs to win
}
\author{
David Flaserstein

David Flaserstein

David Flaserstein
}

