% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propNeeded.R
\name{propNeeded}
\alias{propNeeded}
\title{Shows proportion need to win}
\usage{
propNeeded(object = "Candidate", remainingDelegates = "numeric")
}
\arguments{
\item{object}{of class Candidate}

\item{remainingDelegates}{number of remaining delegates}
}
\value{
print each individual element of Candidate
}
\description{
Shows the porportion of remaining delegates candidate needs to win
}
\author{
David Flaserstein
}

