% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showCandidate.R
\docType{methods}
\name{show,Candidate-method}
\alias{show,Candidate-method}
\title{Show objects of class Candidate}
\usage{
\S4method{show}{Candidate}(object)
}
\arguments{
\item{object}{of class Candidate}
}
\value{
print of class candidate
}
\description{
Inputing an object of class Candidate shows the object
}
\author{
David Flaserstein
}

