% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCandidate.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\alias{calculate_delegates,primaryrace-method}
\alias{initialize,primaryrace-method}
\alias{primaryrace-class}
\title{A candidate's name, party, and delegate count.}
\description{
Object of class \code{Candidate} as created by the \code{createCandidate} function
}
\details{
An object of the class `Candidate' has the following slots:
\itemize{
\item \code{name} The name of the presidential candidate
\item \code{delegatesWon} The delegates the presidential candidate has already won.
\item \code{party} The party of the presidential candidate
\item \code{delegatesNeeded} The amount of delegates the presidential candidate still needs in order to win the nomination
}
}
\author{
Jacob Metz: \email{jacob.metz@wustl.edu}
}

