% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCandidate.R
\docType{methods}
\name{print,Candidate-method}
\alias{print,Candidate-method}
\title{Show objects of class Candidate}
\usage{
\S4method{print}{Candidate}(x, ...)
}
\arguments{
\item{x}{object of class Candidate}
}
\value{
print each individual element of Candidate
}
\description{
Inputing an object of class Candidate shows the object
}
\author{
David Flaserstein
}

